refactor: cookie, swagger
